requirements.txt
allure_python_commons==2.8.13
beautifulsoup4==4.11.1
brotli==1.0.9
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.4
ConfigParser==5.2.0
CouchDB==1.2
cryptography==2.9
gevent==21.12.0
Growl==0.7.1
idna==2.9
Jinja2==2.11.2
JPype1==1.3.0
jsonpath==0.82
kombu==5.2.4
notifo==0.3.0
oauth2==1.9.0.post1
paramiko==2.10.3
pymongo==4.1.0
pynotify==0.1.1
pyOpenSSL==22.0.0
pytest==5.4.1
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
redis==4.2.2
riemann_client==6.5.0
simplejson==3.17.6
slacker==0.14.0
SQLAlchemy==1.4.35
urllib3==1.25.8
